requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.3.0
joblib>=1.3.0
pillow>=8.0.0
tf-nightly

# API dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Visualization
plotly>=5.17.0
folium>=0.15.0
streamlit-folium>=0.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0
